split out HtmlGenerator into 3 mini-libs
